broke out tests, new parm validation test
