+++ tests/pytest_tests/system_tests/tests_launch/test_launch_cli.py
@@ -1,13 +1,8 @@
 import json
-import random
-import subprocess
 
 import pytest
 import wandb
-from wandb.apis.internal import InternalApi
-from wandb.apis.public import Api
 from wandb.cli import cli
-from wandb.sdk.launch.utils import LAUNCH_DEFAULT_PROJECT
 
 REPO_CONST = "test-repo"
 IMAGE_CONST = "fake-image"
@@ -281,208 +276,3 @@ def test_launch_build_with_local(
             "Cannot build a docker image for the resource: local-process"
             in result.output
         )
-
-
-@pytest.mark.parametrize(
-    "image_uri,launch_config",
-    [
-        ("testing111", {}),
-        ("testing222", {"scheduler": {"num_workers": 5}}),
-        ("testing222", {"scheduler": {"num_workers": "5"}}),
-    ],
-    ids=[
-        "working",
-        "num-workers-int",
-        "num-workers-str",
-    ],
-)
-def test_launch_sweep_launch_uri(user, image_uri, launch_config):
-    queue = "testing-" + str(random.random()).replace(".", "")
-    api = InternalApi()
-    public_api = Api()
-    public_api.create_project(LAUNCH_DEFAULT_PROJECT, user)
-
-    # make launch project queue
-    res = api.create_run_queue(
-        entity=user,
-        project=LAUNCH_DEFAULT_PROJECT,
-        queue_name=queue,
-        access="USER",
-    )
-
-    if res.get("success") is not True:
-        raise Exception("create queue" + str(res))
-
-    with open("sweep-config.yaml", "w") as f:
-        json.dump(
-            {
-                "method": "grid",
-                "image_uri": image_uri,
-                "parameters": {"parameter1": {"values": [1, 2, 3]}},
-            },
-            f,
-        )
-
-    out = subprocess.check_output(
-        [
-            "wandb",
-            "launch-sweep",
-            "--sweep_config",
-            "sweep-config.yaml",
-            "-e",
-            user,
-            "-p",
-            LAUNCH_DEFAULT_PROJECT,
-            "-q",
-            queue,
-            "--launch_config",
-            json.dumps(launch_config),
-        ]
-    )
-    wandb.termlog(out)
-
-
-@pytest.mark.parametrize(
-    "image_uri,launch_config,job",
-    [
-        (None, {}, None),
-        ("", {}, None),
-        ("testing111", {"scheduler": {}}, "job123:v1"),
-    ],
-    ids=[
-        "None, empty, None",
-        "empty, None, None",
-        "image + job",
-    ],
-)
-def test_launch_sweep_launch_error(user, image_uri, launch_config, job):
-    queue = "testing-" + str(random.random()).replace(".", "")
-    api = InternalApi()
-    public_api = Api()
-    public_api.create_project(LAUNCH_DEFAULT_PROJECT, user)
-
-    # make launch project queue
-    res = api.create_run_queue(
-        entity=user,
-        project=LAUNCH_DEFAULT_PROJECT,
-        queue_name=queue,
-        access="USER",
-    )
-
-    if not res or res.get("success") is not True:
-        raise Exception("create queue" + str(res))
-
-    with open("sweep-config.yaml", "w") as f:
-        json.dump(
-            {
-                "job": job,
-                "image_uri": image_uri,
-                "method": "grid",
-                "parameters": {"parameter1": {"values": [1, 2, 3]}},
-            },
-            f,
-        )
-
-    out = subprocess.check_output(
-        [
-            "wandb",
-            "launch-sweep",
-            "--sweep_config",
-            "sweep-config.yaml",
-            "-e",
-            user,
-            "-p",
-            LAUNCH_DEFAULT_PROJECT,
-            "-q",
-            queue,
-            "--launch_config",
-            json.dumps(launch_config),
-        ],
-        stderr=subprocess.STDOUT,
-    )
-
-    if job:
-        assert "Sweep has both 'job' and 'image_uri'" in out.decode("utf-8")
-    else:
-        assert "No 'job' nor 'image_uri' found in sweep config" in out.decode("utf-8")
-
-
-def test_launch_sweep_launch_resume(user):
-    api = InternalApi()
-    public_api = Api()
-    public_api.create_project(LAUNCH_DEFAULT_PROJECT, user)
-
-    # make launch project queue
-    res = api.create_run_queue(
-        entity=user,
-        project=LAUNCH_DEFAULT_PROJECT,
-        queue_name="queue",
-        access="USER",
-    )
-
-    if res.get("success") is not True:
-        raise Exception("create queue" + str(res))
-
-    sweep_config = {
-        "job": None,
-        "method": "grid",
-        "image_uri": "test-image:latest",
-        "parameters": {"parameter1": {"values": [1, 2, 3]}},
-    }
-
-    with open("sweep-config.yaml", "w") as f:
-        json.dump(sweep_config, f)
-
-    with pytest.raises(subprocess.CalledProcessError):
-        out = subprocess.check_output(
-            [
-                "wandb",
-                "launch-sweep",
-                "--resume_id",
-                "bogussweepid",
-                "-e",
-                user,
-                "-p",
-                LAUNCH_DEFAULT_PROJECT,
-                "-q",
-                "queue",
-            ],
-            stderr=subprocess.STDOUT,
-        )
-        assert "Launch-sweeps require setting a 'queue'" in out.decode("utf-8")
-
-    # Entity, project, and sweep
-    sweep_id = wandb.sweep(sweep_config, entity=user, project=LAUNCH_DEFAULT_PROJECT)
-
-    # no queue
-    out = subprocess.check_output(
-        [
-            "wandb",
-            "launch-sweep",
-            "--resume_id",
-            sweep_id,
-            "-e",
-            user,
-            "-p",
-            LAUNCH_DEFAULT_PROJECT,
-        ],
-        stderr=subprocess.STDOUT,
-    )
-    assert "Launch-sweeps require setting a 'queue'" in out.decode("utf-8")
-
-    out = subprocess.check_output(
-        [
-            "wandb",
-            "launch-sweep",
-            "--resume_id",
-            sweep_id,
-            "-e",
-            user,
-            "-p",
-            LAUNCH_DEFAULT_PROJECT,
-            "-q",
-            "queue",
-        ],
-        stderr=subprocess.STDOUT,
-    )
-    assert "Scheduler added to launch queue (queue)" in out.decode("utf-8")
